backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.6
httpx==0.27.0
pandas==3.0.2
pandas-ta==0.4.71b0
numpy==2.2.6
apscheduler==3.10.4
google-cloud-aiplatform==1.51.0
vertexai==1.43.0
feedparser==6.0.11
python-dotenv==1.0.1
cryptography==42.0.5
pydantic==2.7.1
pydantic-settings==2.2.1
python-dateutil==2.9.0
pytz==2024.1
groq==0.9.0